% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testes.R
\name{testar_homoscedasticidade}
\alias{testar_homoscedasticidade}
\title{Testar Homocedasticidade}
\usage{
testar_homoscedasticidade(dados, grupos)
}
\arguments{
\item{dados}{Um vetor numérico de dados.}

\item{grupos}{Um vetor de fatores que representam os grupos dos dados.}
}
\value{
Resultados do teste de homocedasticidade.
}
\description{
Esta função realiza testes de homocedasticidade nos dados e fornece recomendações baseadas nos resultados.
}
\examples{
testar_homoscedasticidade(c(1, 2, 3, 4, 5), factor(c(1, 1, 2, 2, 2)))
}
