% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testes.R
\name{testar_normalidade}
\alias{testar_normalidade}
\title{Testar Normalidade}
\usage{
testar_normalidade(dados)
}
\arguments{
\item{dados}{Um vetor numérico de dados.}
}
\value{
Resultados do teste de normalidade.
}
\description{
Esta função realiza testes de normalidade nos dados e fornece recomendações baseadas nos resultados.
}
\examples{
testar_normalidade(c(1, 2, 3, 4, 5))
}
